part2/requirements.txt
boto3>=1.35.0
pyyaml>=6.0
